Now uploading segment flows as hdf5
All NHM streamflows within the DRB watershed are now available in the `drb_seg_outflow_mgd.hdf` file.

This allows for the inflow extraction to be repeated, or for different NHM segments to be used for the Pywr-DRB inflow data without the user needing to use the 92 GB NHM CONUS dataset.